+++ app/src/main/java/com/huynd/skyobserver/models/cheapestflight/month/MonthCheapestResponse.kt
@@ -15,8 +15,8 @@ data class CheapestPricePerMonthResponse(
         @SerializedName("_id")
         val id: ResponseId = ResponseId(),
         @SerializedName("c")
-        val cheapestPrice: Int = 0,
-        var cheapestTotalPrice: Int = 0,
+        val cheapestPrice: Double = 0.0,
+        var cheapestTotalPrice: Double = 0.0,
         @SerializedName("p")
         val carrier: String = ""
 ) : Parcelable
